Implement alert command
The alert command fetches idea information from aha.io and evaluates
whether the ideas with the 'Needs review' flag have been commented
within the last two weeks. If this is not the case, those ideas are
printed to stdout. With the flag -n the result is
posted to a Slacks channel using a webhook url.

Co-authored-by: Robert Gogolok <gogolok@gmail.com>